url: Removes duplicacy while opening actionsheet for images.
Fixes #2249 and #2122. Condenses the code to open Lightbox ActionSheet to a function and lets other files invoke the function without exposing a lot of other functions of lightbox.
